fix casts when linearizing compound assignments
linearize_assignment() emits confusing casts for compound assignments.
Consider p += 1, where p is a pointer.

        oldvalue = cast_pseudo(ep, oldvalue, src->ctype, expr->ctype);
        opcode = opcode_sign(op_trans[expr->op - SPECIAL_BASE], src->ctype);
        dst = add_binary_op(ep, src->ctype, opcode, oldvalue, value);
        value = cast_pseudo(ep, dst, expr->ctype, src->ctype);

In the first call, `oldvalue' is a pointer (expr->ctype), but it is
converted from an integer (src->ctype) to a pointer (expr->ctype);

In the second call, `dst' is an integer (src->ctype), but it is converted
from a pointer (expr->ctype) to an integer (src->ctype).

This patch removes these confusing casts.

Signed-off-by: Xi Wang <xi.wang@gmail.com>