sparse, llvm: fix phi generation
PHI in LLVM is different from that in sparse.  LLVM requires PHI "for
each predecessor basic block of the current block" even if the current
block doesn't use it.  It's tricky to correctly place PHI.

This patch implements a simpler and safer strategy:
1) allocate an alloca for each phi, and
2) translate phi/phisrc to load/store for alloca.
LLVM optimizations will promote load/store to registers.

Cc: Pekka Enberg <penberg@kernel.org>
Cc: Christopher Li <sparse@chrisli.org>
Cc: Jeff Garzik <jgarzik@redhat.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Jonathan Neuschäfer <j.neuschaefer@gmx.net>
Signed-off-by: Xi Wang <xi.wang@gmail.com>